Added data link to execution working directory (#3)
This commit adds a soft link to the input data of a step in its working
directory. Before this commit, the function `_exec` had a parameter,
`InitialWorkDir`, to decide whenever create the soft link. Now, it has
been removed and the creation of the input data soft link is the standard
behavior.