Added `Workflow` output ports (#2)
This commit adds output ports of the entire workflow. When a
DAX workflow is translated, a new step is added for each workflow
output port. These steps are at the end of a trace and they manage
the copy data from `workdir` to `outdir` in a location.

Moreover, this commit adds the `outdir` attribute in the `Location`
class. The data of the workflow output ports are copied in the `outdir`
path of the related location.  The `workdir` attribute of the `Location`
class instead is used as scratch directory.